requirements.txt
django
psycopj2
djangorestframework